fix: Rename parser adapters
Fixes #1847
